Move some common helper functions to util.py
